Adding KML data broken into three rank files
Paraphrased from Dr. Witschey:
Ranks are a reflection of site size. In this data set they are largely
based upon the research of Garza and Kurjack in the 80s.
Size is a rough measure of areal extent; amount/size of monumental
architecture and population...but it is a very fuzzy measure.

Paraphrased from Dr. Brown
Rank is probably best thought of as a measure
of sociocultural complexity applied to urban form. It considers overall
site size, volume of monumental (civic-ceremonial or public)
construction, and the functional diversity of the architecture (i.e.,
different types of buildings, such as markets, palaces, temples, as well
as simple homes). There are, however, a number of different ranking
systems, and since we rarely have full information about a single site,
they are generally applied impressionistically. So, I Garza and Kurjack
used 4 ranks but other authors have used five or more. I believe the
base map we used for Guatemala employed a system with only three
ranks.))