First pass at commas to semicolons
I'm getting some bad data in the PostGIS database because of the quoted
fields where commas are used within a delimited field. So I'm switching
these non-delimiting commas to semi-colons. This is the first pass
before taking another loop with the regex to get the 2nd set of commas
within the delimited fields.